Frontend kira

+++ app_server/src/storages/mongo/repositories/task.py
@@ -1,5 +1,6 @@
 from beanie import PydanticObjectId
 from src.storages.mongo.models.task import Task, TaskCreate, TaskUpdate
+from src.storages.mongo.repositories.document import document_repository
 
 
 class TaskRepository:
@@ -20,10 +21,15 @@ class TaskRepository:
 
         return await task.set(task_update.model_dump())
 
-    async def delete(self, task_id: PydanticObjectId) -> Task | None:
+    async def delete(self, task_id: PydanticObjectId):
         task = await Task.find_one({"_id": task_id})
         if task is None:
             return None
+        document = await document_repository.read(task.document_id)
+        if document is None:
+            return None
+        document.tasks = [t for t in document.tasks if t != task_id]
+        await document_repository.update(document.id, document)
 
         return await task.delete()
